+++ src/sendquery.js
@@ -16,8 +16,8 @@ import { auth } from './auth.js';
 
 import {
   cleanHeaderParams,
-  cleanQuery,
   getHeaderParams,
+  getTrailingParams,
   loadQuery,
   resolveParameterDiff,
 } from './util.js';
@@ -32,17 +32,19 @@ async function processParams(query, params) {
   const rawQuery = await loadQuery(query);
   const headerParams = getHeaderParams(rawQuery);
   const description = headerParams.description || '';
-  const loadedQuery = cleanQuery(rawQuery);
+  const loadedQuery = rawQuery;
   const requestParams = resolveParameterDiff(
     cleanHeaderParams(loadedQuery, params),
     cleanHeaderParams(loadedQuery, headerParams),
   );
+  const responseDetails = getTrailingParams(loadedQuery);
 
   return {
     headerParams,
     description,
     loadedQuery,
     requestParams,
+    responseDetails,
   };
 }
 
@@ -62,6 +64,7 @@ export async function execute(email, key, project, query, _, params = {}) {
     description,
     loadedQuery,
     requestParams,
+    responseDetails,
   } = await processParams(query, params);
   try {
     const credentials = await auth(email, key.replace(/\\n/g, '\n'));
@@ -91,12 +94,28 @@ export async function execute(email, key, project, query, _, params = {}) {
         return true;
       };
 
+      let stream;
+      let rootJob;
       const q = loadedQuery;
-      const stream = await bq.createQueryStream({
-        query: q,
-        maxResults: params.limit,
-        params: requestParams,
-      });
+
+      const responseMetadata = {};
+      if (loadedQuery.indexOf('# hlx:metadata') > -1) {
+        const jobs = await bq.createQueryJob({
+          query: q,
+          params: requestParams,
+        });
+        stream = await jobs[0].getQueryResultsStream();
+
+        // we have multiple jobs, so we need to inspect the first job
+        // to get the list of all jobs.
+        [rootJob] = jobs;
+      } else {
+        stream = await bq.createQueryStream({
+          query: q,
+          maxResults: params.limit,
+          params: requestParams,
+        });
+      }
       stream
         .on('data', (row) => (spaceleft() ? results.push(row) : resolve({
           headers,
@@ -104,6 +123,8 @@ export async function execute(email, key, project, query, _, params = {}) {
           results,
           description,
           requestParams,
+          responseDetails,
+          responseMetadata,
         })))
         .on(
           'error',
@@ -113,12 +134,27 @@ export async function execute(email, key, project, query, _, params = {}) {
           },
         )
         .on('end', async () => {
+          if (rootJob) {
+            // try to get the list of child jobs. We need to wait until the
+            // root job has finished, otherwise the list of child jobs is not
+            // complete.
+            const [childJobs] = await bq.getJobs({
+              parentJobId: rootJob.metadata.jobReference.jobId,
+            });
+            const metadata = childJobs[1]; // jobs are ordered in descending order by execution time
+            if (metadata) {
+              const [metadataResults] = await metadata.getQueryResults();
+              responseMetadata.totalRows = metadataResults[0]?.total_rows;
+            }
+          }
           resolve({
             headers,
             truncated: false,
             results,
             description,
             requestParams,
+            responseDetails,
+            responseMetadata,
           });
         });
     });