+++ test/testUtils.js
@@ -13,17 +13,37 @@
 /* eslint-env mocha */
 import assert from 'assert';
 import {
+  chartify,
   cleanHeaderParams,
-  cleanQuery, cleanRequestParams, csvify,
+  cleanRequestParams, csvify,
   getHeaderParams,
+  getTrailingParams,
   loadQuery, resolveParameterDiff,
   sshonify,
+  validParamCheck, extractQueryPath,
 } from '../src/util.js';
 
 describe('testing util functions', () => {
+  it('extractQueryPath parses a run-query path to find the folder/file for a query', async () => {
+    assert.equal(extractQueryPath('/helix-services/run-query/file'), 'file');
+    assert.equal(extractQueryPath('/helix-services/run-query@v1/file'), 'file');
+    assert.equal(extractQueryPath('/helix-services/run-query@v2/folder/file'), 'folder/file');
+    assert.equal(extractQueryPath('/helix-services/run-query@ci123/file'), 'file');
+    assert.equal(extractQueryPath('/helix-services/run-query@ci456/folder/file'), 'folder/file');
+    assert.equal(extractQueryPath('/helix-services/run-query/ci789/file'), 'file');
+    assert.equal(extractQueryPath('/helix-services/run-query/ci123/folder/file'), 'folder/file');
+    assert.equal(extractQueryPath('/helix-services/run-query@v3/folder/folder/file'), 'folder/folder/file');
+    assert.equal(extractQueryPath('/helix-services/run-query/3.3.0/folder/folder/file'), 'folder/folder/file');
+  });
+
   it('loadQuery loads a query', async () => {
     const result = await loadQuery('rum-dashboard');
     assert.ok(result.match(/select/i));
+  });
+
+  it('loadQuery works with trailing parameters', async () => {
+    const result = await loadQuery('rum-dashboard');
+    assert.equal(Object.keys(getTrailingParams(result)).length, 31);
   });
 
   it('loadQuery throws with bad query file', async () => {
@@ -42,34 +62,6 @@ describe('testing util functions', () => {
     const EXPECTED = { 'helix-param': 'helix', 'helix-param2': 'helix2', 'helix-param3': 'helix3' };
     const ACTUAL = getHeaderParams(fakeQuery);
     assert.deepEqual(EXPECTED, ACTUAL);
-  });
-
-  it('query parameters are cleaned from query', () => {
-    const fakeQuery = `--- helix-param: helix
---- helix-param2: helix2
---- helix-param3: helix3
-#This is A random Comment
-SELECT req_url, count(req_http_X_CDN_Request_ID) AS visits, resp_http_Content_Type, status_code
-    FROM ^tablename
-    WHERE 
-      resp_http_Content_Type LIKE "text/html%" AND
-      status_code LIKE "404"
-    GROUP BY
-      req_url, resp_http_Content_Type, status_code 
-    ORDER BY visits DESC
-    LIMIT @limit`;
-
-    const EXPECTED = `SELECT req_url, count(req_http_X_CDN_Request_ID) AS visits, resp_http_Content_Type, status_code
-    FROM ^tablename
-    WHERE 
-      resp_http_Content_Type LIKE "text/html%" AND
-      status_code LIKE "404"
-    GROUP BY
-      req_url, resp_http_Content_Type, status_code 
-    ORDER BY visits DESC
-    LIMIT @limit`;
-    const ACTUAL = cleanQuery(fakeQuery);
-    assert.equal(EXPECTED, ACTUAL);
   });
 
   it('resolveParameterDiff fills in empty params with defaults', () => {
@@ -202,6 +194,44 @@ SELECT @something1, @something2 WHERE @tablename`;
     const expected = '';
     assert.equal(result, expected);
   });
+
+  it('validParamCheck returns false if there is array in param', () => {
+    const badParams = {
+      one: 'okay',
+      two: 'okay',
+      three: [1, 2, 3, 4, 5],
+      four: 'okay',
+    };
+
+    const EXPECTED_STATUS = '400';
+    const EXPECTED_MESSAGE = 'Duplicate URL parameters found';
+
+    try {
+      validParamCheck(badParams);
+    } catch (e) {
+      assert.deepEqual(EXPECTED_STATUS, e.statusCode);
+      assert.deepEqual(EXPECTED_MESSAGE, e.message);
+    }
+  });
+
+  it('resolveParameterDiff will throw error with 400 if array detected', () => {
+    const badParams = {
+      one: 'okay',
+      two: 'okay',
+      three: [1, 2, 3, 4, 5],
+      four: 'okay',
+    };
+
+    const EXPECTED_STATUS = 400;
+    const EXPECTED_MESSAGE = 'Duplicate URL parameters found';
+
+    try {
+      resolveParameterDiff(badParams, {});
+    } catch (e) {
+      assert.deepEqual(EXPECTED_STATUS, e.statusCode);
+      assert.deepEqual(EXPECTED_MESSAGE, e.message);
+    }
+  });
 });
 
 describe('Test SSHONify', () => {
@@ -272,9 +302,26 @@ describe('Test SSHONify', () => {
       requestParams: {
         limit: 30, interval: 1, offset: '0', url: 'blog.adobe.com', checkpoint: 'viewblock', source: '-',
       },
+      responseDetails: {
+        checkpoint: 'name of the checkpoint, i.e. the event in the page load or interaction sequence that was observed.',
+        source: 'CSS id or class name of the element that triggered the checkpoint.',
+        ids: 'number of unique RUM ids that triggered the checkpoint.',
+        pages: 'number of unique pages that triggered the checkpoint.',
+        topurl: 'most frequently observed URL that triggered the checkpoint.',
+        views: 'interpolated number of pageviews that triggered the checkpoint.',
+        actions: 'number of times the checkpoint was triggered. This may be greater than the number of unique ids if the same id triggered the checkpoint multiple times.',
+        actions_per_view: 'average number of times the checkpoint was triggered per pageview.',
+      },
       truncated: false,
     };
-    const sshon = sshonify(input.results, input.description, input.requestParams, input.truncated);
+    const sshon = sshonify(
+      input.results,
+      input.description,
+      input.requestParams,
+      input.responseDetails,
+      {},
+      input.truncated,
+    );
     assert.deepStrictEqual(
       JSON.parse(sshon),
       {
@@ -321,6 +368,46 @@ describe('Test SSHONify', () => {
               type: 'request parameter',
               value: '-',
             },
+            {
+              name: 'checkpoint',
+              type: 'response detail',
+              value: 'name of the checkpoint, i.e. the event in the page load or interaction sequence that was observed.',
+            },
+            {
+              name: 'source',
+              type: 'response detail',
+              value: 'CSS id or class name of the element that triggered the checkpoint.',
+            },
+            {
+              name: 'ids',
+              type: 'response detail',
+              value: 'number of unique RUM ids that triggered the checkpoint.',
+            },
+            {
+              name: 'pages',
+              type: 'response detail',
+              value: 'number of unique pages that triggered the checkpoint.',
+            },
+            {
+              name: 'topurl',
+              type: 'response detail',
+              value: 'most frequently observed URL that triggered the checkpoint.',
+            },
+            {
+              name: 'views',
+              type: 'response detail',
+              value: 'interpolated number of pageviews that triggered the checkpoint.',
+            },
+            {
+              name: 'actions',
+              type: 'response detail',
+              value: 'number of times the checkpoint was triggered. This may be greater than the number of unique ids if the same id triggered the checkpoint multiple times.',
+            },
+            {
+              name: 'actions_per_view',
+              type: 'response detail',
+              value: 'average number of times the checkpoint was triggered per pageview.',
+            },
           ],
           columns: [
             'name',
@@ -333,8 +420,8 @@ describe('Test SSHONify', () => {
         },
         results: {
           limit: 30,
-          offset: '0',
-          total: '0',
+          offset: 0,
+          total: 30,
           columns: [
             'checkpoint',
             'source',
@@ -422,7 +509,14 @@ describe('Test SSHONify', () => {
       },
       truncated: false,
     };
-    const sshon = sshonify(input.results, input.description, input.requestParams, input.truncated);
+    const sshon = sshonify(
+      input.results,
+      input.description,
+      input.requestParams,
+      {},
+      {},
+      input.truncated,
+    );
     assert.deepStrictEqual(
       JSON.parse(sshon),
       {
@@ -481,13 +575,291 @@ describe('Test SSHONify', () => {
         },
         results: {
           limit: 30,
-          offset: '0',
-          total: '0',
+          offset: 0,
+          total: 0,
           columns: [],
           data: [],
         },
       },
 
     );
+  });
+});
+
+describe('Test chartify', () => {
+  it('chartify generates chart', () => {
+    const chartconfig = {
+      type: 'horizontalBar',
+      data: {
+        labels: '@source',
+        datasets: [
+          {
+            label: 'Views',
+            backgroundColor: 'rgba(255, 99, 132, 0.5)',
+            borderColor: 'rgb(255, 99, 132)',
+            borderWidth: 1,
+            data: '@views',
+          },
+        ],
+      },
+      options: {
+        elements: {
+          rectangle: {
+            borderWidth: 2,
+          },
+        },
+        responsive: true,
+        legend: {
+          position: 'right',
+        },
+        title: {
+          display: true,
+          text: 'Most popular Blocks by Views',
+        },
+      },
+    };
+    const input = {
+      results: [{
+        checkpoint: 'viewblock', source: '.article-header', ids: 3258, pages: 1031, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '32580', actions: '32580', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.images', ids: 2100, pages: 651, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '21000', actions: '21000', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.banner', ids: 1350, pages: 294, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '13500', actions: '13500', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.tags', ids: 1311, pages: 383, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '13110', actions: '13110', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.recommended-articles', ids: 1117, pages: 271, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '11170', actions: '11170', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.embed', ids: 769, pages: 226, topurl: 'https://blog.adobe.com/en/publish/2023/04/18/new-adobe-lightroom-ai-innovations-empower-everyone-edit-like-pro', views: '7690', actions: '7690', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.animation', ids: 473, pages: 55, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '4730', actions: '4730', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.featured-article', ids: 444, pages: 36, topurl: 'https://blog.adobe.com/', views: '4440', actions: '4440', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.video', ids: 420, pages: 24, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '4200', actions: '4200', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.article-feed', ids: 359, pages: 105, topurl: 'https://blog.adobe.com/en/topics/spark', views: '3590', actions: '3590', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.tag-header', ids: 222, pages: 63, topurl: 'https://blog.adobe.com/en/topics/spark', views: '2220', actions: '2220', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.marquee', ids: 179, pages: 55, topurl: 'https://blog.adobe.com/jp/publish/2023/03/23/cc-video-premierepro-interview-guts100tv?trackingid=RYGDMVVQ&mv=email', views: '1790', actions: '1790', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.table-of-contents', ids: 153, pages: 66, topurl: 'https://blog.adobe.com/en/publish/2023/03/21/responsible-innovation-age-of-generative-ai', views: '1530', actions: '1530', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.error-attr', ids: 65, pages: 44, topurl: 'https://blog.adobe.com/en/2019/05/30/the-future-of-adobe-air', views: '650', actions: '650', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.error-text', ids: 65, pages: 44, topurl: 'https://blog.adobe.com/en/2019/05/30/the-future-of-adobe-air', views: '650', actions: '650', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.pull-quote', ids: 47, pages: 27, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/national-volunteer-week', views: '470', actions: '470', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.author-header', ids: 38, pages: 28, topurl: 'https://blog.adobe.com/en/authors/adobe-communications-team', views: '380', actions: '380', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.toolkit', ids: 27, pages: 5, topurl: 'https://blog.adobe.com/en/topics/adobe-life', views: '270', actions: '270', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.newsletter-modal', ids: 26, pages: 16, topurl: 'https://blog.adobe.com/en/2019/05/30/the-future-of-adobe-air', views: '260', actions: '260', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.image', ids: 20, pages: 7, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/national-volunteer-week', views: '200', actions: '200', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '#tabs-0', ids: 15, pages: 1, topurl: 'https://blog.adobe.com/en/topics/adobe-life', views: '150', actions: '150', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.infographic', ids: 14, pages: 11, topurl: 'https://blog.adobe.com/en/publish/2023/03/23/practicing-digital-accessibility-in-workplace1', views: '140', actions: '140', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.columns', ids: 14, pages: 2, topurl: 'https://blog.adobe.com/en/topics/adobe-life', views: '140', actions: '140', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.infograph', ids: 10, pages: 4, topurl: 'https://blog.adobe.com/en/publish/2022/07/27/day-in-the-life-of-a-3d-designer', views: '100', actions: '100', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.art-culos-recomendados', ids: 5, pages: 1, topurl: 'https://blog.adobe.com/es/publish/2023/04/17/reimaginando-video-audio-adobe-firefly', views: '50', actions: '50', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.social-links', ids: 5, pages: 5, topurl: 'https://blog.adobe.com/en/authors/cooper-savage', views: '50', actions: '50', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.carousel', ids: 3, pages: 1, topurl: 'https://blog.adobe.com/en/publish/2023/03/08/expanding-adobe-presence-commitments-san-jose', views: '30', actions: '30', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.namecard', ids: 3, pages: 1, topurl: 'https://blog.adobe.com/jp/publish/2018/04/26/dtp-illustrator-kihon-tips-10', views: '30', actions: '30', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.image-50', ids: 2, pages: 1, topurl: 'https://blog.adobe.com/es/publish/2021/11/24/tarjetas-de-cumpleanos-disenadas-por-heylovelygirl-para-descargar-gratis', views: '20', actions: '20', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.full-width-banner', ids: 2, pages: 1, topurl: 'https://blog.adobe.com/en/publish/2023/02/24/academy-award-recognizes-substance-3d-for-transforming-movie-visual-effect-animation', views: '20', actions: '20', actions_per_view: '1',
+      }],
+      description: 'Get popularity data for RUM source attribute values, filtered by checkpoint',
+      requestParams: {
+        limit: 30,
+        interval: 1,
+        offset: '0',
+        url: 'blog.adobe.com',
+        checkpoint: 'viewblock',
+        source: '-',
+        chart: JSON.stringify(chartconfig),
+      },
+      responseDetails: {
+        checkpoint: 'name of the checkpoint, i.e. the event in the page load or interaction sequence that was observed.',
+        source: 'CSS id or class name of the element that triggered the checkpoint.',
+        ids: 'number of unique RUM ids that triggered the checkpoint.',
+        pages: 'number of unique pages that triggered the checkpoint.',
+        topurl: 'most frequently observed URL that triggered the checkpoint.',
+        views: 'interpolated number of pageviews that triggered the checkpoint.',
+        actions: 'number of times the checkpoint was triggered. This may be greater than the number of unique ids if the same id triggered the checkpoint multiple times.',
+        actions_per_view: 'average number of times the checkpoint was triggered per pageview.',
+      },
+      truncated: false,
+    };
+    const chart = chartify(
+      input.results,
+      input.description,
+      input.requestParams,
+    );
+    assert(chart);
+    assert.equal(chart, '{"type":"horizontalBar","data":{"labels":[".article-header",".images",".banner",".tags",".recommended-articles",".embed",".animation",".featured-article",".video",".article-feed",".tag-header",".marquee",".table-of-contents",".error-attr",".error-text",".pull-quote",".author-header",".toolkit",".newsletter-modal",".image","#tabs-0",".infographic",".columns",".infograph",".art-culos-recomendados",".social-links",".carousel",".namecard",".image-50",".full-width-banner"],"datasets":[{"label":"Views","backgroundColor":"rgba(255, 99, 132, 0.5)","borderColor":"rgb(255, 99, 132)","borderWidth":1,"data":["32580","21000","13500","13110","11170","7690","4730","4440","4200","3590","2220","1790","1530","650","650","470","380","270","260","200","150","140","140","100","50","50","30","30","20","20"]}]},"options":{"elements":{"rectangle":{"borderWidth":2}},"responsive":true,"legend":{"position":"right"},"title":{"display":true,"text":"Most popular Blocks by Views"}}}');
+  });
+
+  it('chartify generates chart from string config', () => {
+    const chartconfig = `{
+      type: 'horizontalBar',
+      data: {
+        labels: @source,
+        datasets: [
+          {
+            label: 'Views',
+            backgroundColor: getGradientFillHelper('vertical', ['#eb3639', '#a336eb', '#36a2eb']),
+            borderWidth: 1,
+            data: @pages,
+          },
+        ],
+      },
+      options: {
+        elements: {
+          rectangle: {
+            borderWidth: 2,
+          },
+        },
+        responsive: true,
+        legend: {
+          position: 'right',
+        },
+        title: {
+          display: true,
+          text: 'Most popular Blocks by number of pages on which they were viewed',
+        },
+      },
+    }`;
+    const input = {
+      results: [{
+        checkpoint: 'viewblock', source: '.article-header', ids: 3258, pages: 1031, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '32580', actions: '32580', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.images', ids: 2100, pages: 651, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '21000', actions: '21000', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.banner', ids: 1350, pages: 294, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '13500', actions: '13500', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.tags', ids: 1311, pages: 383, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '13110', actions: '13110', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.recommended-articles', ids: 1117, pages: 271, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '11170', actions: '11170', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.embed', ids: 769, pages: 226, topurl: 'https://blog.adobe.com/en/publish/2023/04/18/new-adobe-lightroom-ai-innovations-empower-everyone-edit-like-pro', views: '7690', actions: '7690', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.animation', ids: 473, pages: 55, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '4730', actions: '4730', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.featured-article', ids: 444, pages: 36, topurl: 'https://blog.adobe.com/', views: '4440', actions: '4440', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.video', ids: 420, pages: 24, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/reimagining-video-audio-adobe-firefly', views: '4200', actions: '4200', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.article-feed', ids: 359, pages: 105, topurl: 'https://blog.adobe.com/en/topics/spark', views: '3590', actions: '3590', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.tag-header', ids: 222, pages: 63, topurl: 'https://blog.adobe.com/en/topics/spark', views: '2220', actions: '2220', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.marquee', ids: 179, pages: 55, topurl: 'https://blog.adobe.com/jp/publish/2023/03/23/cc-video-premierepro-interview-guts100tv?trackingid=RYGDMVVQ&mv=email', views: '1790', actions: '1790', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.table-of-contents', ids: 153, pages: 66, topurl: 'https://blog.adobe.com/en/publish/2023/03/21/responsible-innovation-age-of-generative-ai', views: '1530', actions: '1530', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.error-attr', ids: 65, pages: 44, topurl: 'https://blog.adobe.com/en/2019/05/30/the-future-of-adobe-air', views: '650', actions: '650', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.error-text', ids: 65, pages: 44, topurl: 'https://blog.adobe.com/en/2019/05/30/the-future-of-adobe-air', views: '650', actions: '650', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.pull-quote', ids: 47, pages: 27, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/national-volunteer-week', views: '470', actions: '470', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.author-header', ids: 38, pages: 28, topurl: 'https://blog.adobe.com/en/authors/adobe-communications-team', views: '380', actions: '380', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.toolkit', ids: 27, pages: 5, topurl: 'https://blog.adobe.com/en/topics/adobe-life', views: '270', actions: '270', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.newsletter-modal', ids: 26, pages: 16, topurl: 'https://blog.adobe.com/en/2019/05/30/the-future-of-adobe-air', views: '260', actions: '260', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.image', ids: 20, pages: 7, topurl: 'https://blog.adobe.com/en/publish/2023/04/17/national-volunteer-week', views: '200', actions: '200', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '#tabs-0', ids: 15, pages: 1, topurl: 'https://blog.adobe.com/en/topics/adobe-life', views: '150', actions: '150', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.infographic', ids: 14, pages: 11, topurl: 'https://blog.adobe.com/en/publish/2023/03/23/practicing-digital-accessibility-in-workplace1', views: '140', actions: '140', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.columns', ids: 14, pages: 2, topurl: 'https://blog.adobe.com/en/topics/adobe-life', views: '140', actions: '140', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.infograph', ids: 10, pages: 4, topurl: 'https://blog.adobe.com/en/publish/2022/07/27/day-in-the-life-of-a-3d-designer', views: '100', actions: '100', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.art-culos-recomendados', ids: 5, pages: 1, topurl: 'https://blog.adobe.com/es/publish/2023/04/17/reimaginando-video-audio-adobe-firefly', views: '50', actions: '50', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.social-links', ids: 5, pages: 5, topurl: 'https://blog.adobe.com/en/authors/cooper-savage', views: '50', actions: '50', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.carousel', ids: 3, pages: 1, topurl: 'https://blog.adobe.com/en/publish/2023/03/08/expanding-adobe-presence-commitments-san-jose', views: '30', actions: '30', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.namecard', ids: 3, pages: 1, topurl: 'https://blog.adobe.com/jp/publish/2018/04/26/dtp-illustrator-kihon-tips-10', views: '30', actions: '30', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.image-50', ids: 2, pages: 1, topurl: 'https://blog.adobe.com/es/publish/2021/11/24/tarjetas-de-cumpleanos-disenadas-por-heylovelygirl-para-descargar-gratis', views: '20', actions: '20', actions_per_view: '1',
+      }, {
+        checkpoint: 'viewblock', source: '.full-width-banner', ids: 2, pages: 1, topurl: 'https://blog.adobe.com/en/publish/2023/02/24/academy-award-recognizes-substance-3d-for-transforming-movie-visual-effect-animation', views: '20', actions: '20', actions_per_view: '1',
+      }],
+      description: 'Get popularity data for RUM source attribute values, filtered by checkpoint',
+      requestParams: {
+        limit: 30,
+        interval: 1,
+        offset: '0',
+        url: 'blog.adobe.com',
+        checkpoint: 'viewblock',
+        source: '-',
+        chart: chartconfig,
+      },
+      responseDetails: {
+        checkpoint: 'name of the checkpoint, i.e. the event in the page load or interaction sequence that was observed.',
+        source: 'CSS id or class name of the element that triggered the checkpoint.',
+        ids: 'number of unique RUM ids that triggered the checkpoint.',
+        pages: 'number of unique pages that triggered the checkpoint.',
+        topurl: 'most frequently observed URL that triggered the checkpoint.',
+        views: 'interpolated number of pageviews that triggered the checkpoint.',
+        actions: 'number of times the checkpoint was triggered. This may be greater than the number of unique ids if the same id triggered the checkpoint multiple times.',
+        actions_per_view: 'average number of times the checkpoint was triggered per pageview.',
+      },
+      truncated: false,
+    };
+    const chart = chartify(
+      input.results,
+      input.description,
+      input.requestParams,
+    );
+    assert(chart);
+    const expected = `{
+      type: 'horizontalBar',
+      data: {
+        labels: [".article-header",".images",".banner",".tags",".recommended-articles",".embed",".animation",".featured-article",".video",".article-feed",".tag-header",".marquee",".table-of-contents",".error-attr",".error-text",".pull-quote",".author-header",".toolkit",".newsletter-modal",".image","#tabs-0",".infographic",".columns",".infograph",".art-culos-recomendados",".social-links",".carousel",".namecard",".image-50",".full-width-banner"],
+        datasets: [
+          {
+            label: 'Views',
+            backgroundColor: getGradientFillHelper('vertical', ['#eb3639', '#a336eb', '#36a2eb']),
+            borderWidth: 1,
+            data: [1031,651,294,383,271,226,55,36,24,105,63,55,66,44,44,27,28,5,16,7,1,11,2,4,1,5,1,1,1,1],
+          },
+        ],
+      },
+      options: {
+        elements: {
+          rectangle: {
+            borderWidth: 2,
+          },
+        },
+        responsive: true,
+        legend: {
+          position: 'right',
+        },
+        title: {
+          display: true,
+          text: 'Most popular Blocks by number of pages on which they were viewed',
+        },
+      },
+    }`;
+    assert.equal(chart, expected);
   });
 });